fix(landingPage): fix media queries to eliminate text overlaying background img on small screen sizes, updated favicon, renamed files to jsx as appropriate
