+++ config-overrides.js
@@ -1,16 +1,16 @@
 const { override, fixBabelImports, addLessLoader } = require('customize-cra');
 
 module.exports = override(
-	fixBabelImports('import', {
-		libraryName: 'antd',
-		libraryDirectory: 'es',
-		style: true
-	}),
-	addLessLoader({
-		javascriptEnabled: true,
-		modifyVars: {
-			'@primary-color': '#BB1333',
-			'@font-family': "'Nunito', sans-serif"
-		}
-	})
+  fixBabelImports('import', {
+    libraryName: 'antd',
+    libraryDirectory: 'es',
+    style: true,
+  }),
+  addLessLoader({
+    javascriptEnabled: true,
+    modifyVars: {
+      '@primary-color': '#BB1333',
+      '@font-family': "'Nunito', sans-serif",
+    },
+  })
 );
