+++ src/App.jsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import AppRouter from './components/Router';
 
-
 const App = () => {
-  return (
-    <AppRouter />
-  );
-}
+  return <AppRouter />;
+};
 
 export default App;
