--- src/components/Layout/DashboardLayout.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Route } from 'react-router-dom';
-
-
-const DashboardLayout = ({component: Component, ...rest}) => {
-    return (
-        <Route {...rest} render={props => (
-            <Component {...props} />
-        )}/>
-    )
-};
-
-export default DashboardLayout;
