+++ src/components/Router/index.jsx
@@ -4,11 +4,11 @@ import DashboardLayout from '../Layout/DashboardLayout';
 import Home from '../../pages/Home';
 
 const AppRouter = () => {
-    return (
-        <BrowserRouter>
-            <DashboardLayout exact path="/" component={Home} />
-        </BrowserRouter>
-    )
-}
+  return (
+    <BrowserRouter>
+      <DashboardLayout exact path="/" component={Home} />
+    </BrowserRouter>
+  );
+};
 
 export default AppRouter;
